mc
git-svn-id: http://svn.code.sf.net/p/lejos/code/trunk@3136 06c8f799-c779-4783-8af7-179afedb67c8

+++ classes/java/lang/Math.java
@@ -502,8 +502,7 @@ public final class Math
 	{
 		if ((a < -1) || (a > 1))
 		{
-			return Float.NaN; // TODO: We don't have Double.NaN yet, so using
-			// Float for now.
+			return Double.NaN;
 		}
 		return PI / 2 - atan(a / sqrt(1 - a * a));
 	}
